Class 158: Add ScotsRail skin
Skin the train with a ScotRail skin. This is fairly invasive. We now
split the train into 4 double window sections and a single window
section in the middle, and separate the tables into a separate file. The
bogie is moved to a different item damage to make space for the 5 body
parts.

Signed-off-by: James Hogan <james@albanarts.com>