Attempt to reduce lag
Turn off misc and train collisions, and reduce wait distance to 10
blocks to reduce the track that must be kept track of. Note, this will
result in trains being closer together and potentially getting stuck if
you haven't set up mutexes in the appropriate places.

Also tweak the delay down to 5 seconds so that trains start up again a
bit quicker.

Signed-off-by: James Hogan <james@albanarts.com>