Generate item JSON from CSV for 1.21.4
Minecraft 1.21.4 requires item JSON files to be in a different place and
in a different format. Therefore condense the information in
golden_pickaxe.json into a csv file, and make the build system generate
both old and new json files to place in the resource pack.

Update the pack format to 46 to reflect the changes.

Fixes #7
Signed-off-by: James Hogan <james@albanarts.com>